[IMP] Improves language handling
Handles language preferences more robustly.

It now checks for and respects `force_app_lang`, sending a header if necessary to enforce the language. If the language matches the forced language, it disables the `force_app_lang` setting. If no forced language, the code will write the new language to the partner.